backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
requests
python-multipart
python-dotenv
fastapi-mail
httpx
